Update tests and move Debugger type to models
