+++ src/models/types.ts
@@ -1,0 +1,4 @@
+export type LeptonDebugger = {
+  log: (msg: string) => void;
+  error: (error: Error) => void;
+}
